Changed appropriate ints to doubles

+++ a5/a5/Creature.cpp
@@ -11,7 +11,7 @@ Creature::~Creature() {
 
 }
 
-void Creature::loseHp(int damage) {
+void Creature::loseHp(double damage) {
 	hp -= damage;
 	if (hp <= 0) {
 		hp = 0;
@@ -19,11 +19,11 @@ void Creature::loseHp(int damage) {
 	}
 }
 
-int Creature::getHp() {
+double Creature::getHp() {
 	return hp;
 }
 
-int Creature::getDefense() {
+double Creature::getDefense() {
 	return def;
 }
